django_siddhi/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
